+++ Popup/settings.js
@@ -3,6 +3,7 @@ function init() {
 
     document.getElementById("btnVal").onclick = changeBookmarks;
     document.getElementById("cbActive").onchange = changeActiveState;
+    document.getElementById("cbActiveDiscord").onchange = changeActiveDiscordState;
     document.getElementById("btnCacheDelete").onclick = deleteCache;
     document.getElementById("btnUnauthorize").onclick = unauthorize;
 
@@ -11,6 +12,9 @@ function init() {
     })
     getBookmarkFolderName(name => {
         document.getElementById("tbBookmarks").value = name;
+    })
+    getActiveDiscordState(active => {
+        document.getElementById("cbActiveDiscord").checked = active;
     })
 }
 
@@ -32,6 +36,16 @@ function getBookmarkFolderName(callb) {
     });
 }
 
+function getActiveDiscordState(callb) {
+    chrome.storage.local.get("MAL_Settings_DiscordActive", res => {
+        if (res.MAL_Settings_DiscordActive != "")
+            callb(res.MAL_Settings_DiscordActive);
+        else
+            callb(false);
+    });
+}
+
+
 function changeBookmarks() {
     chrome.storage.local.set({ "MAL_Settings_Bookmarks": document.getElementById("tbBookmarks").value }, function () {
         chrome.runtime.sendMessage(
@@ -49,6 +63,17 @@ function changeActiveState(event) {
         chrome.runtime.sendMessage(
             {
                 type: "CHANGED_ACTIVE",
+                value: event.target.checked
+            },
+            () => { }
+        );
+    });
+}
+function changeActiveDiscordState(event) {
+    chrome.storage.local.set({ "MAL_Settings_DiscordActive": event.target.checked }, function () {
+        chrome.runtime.sendMessage(
+            {
+                type: "CHANGED_ACTIVE_DISCORD",
                 value: event.target.checked
             },
             () => { }